Adding automatic generation of an index for the Documentation that lists each array by its default name and what filters create that array. Cleaning up compiler warnings.

+++ Source/DREAM3DLib/ProcessingFilters/IdentifySample.cpp
@@ -191,7 +191,7 @@ void IdentifySample::execute()
         }
         count++;
       }
-      if(currentvlist.size() >= biggestBlock)
+      if(static_cast<int>(currentvlist.size()) >= biggestBlock)
       {
         biggestBlock = currentvlist.size();
         for(int j = 0; j < totalPoints; j++)
@@ -246,7 +246,7 @@ void IdentifySample::execute()
         }
         count++;
       }
-      if(currentvlist.size() >= biggestBlock)
+      if(static_cast<int>(currentvlist.size()) >= biggestBlock)
       {
         biggestBlock = currentvlist.size();
         for(int j = 0; j < totalPoints; j++)
